Refactor character page into modular dashboard
